Avoid duplicate artifacts
To avoid duplicating artifacts in the manifest, each artifact is resolved to the
NixArtifact type before being stored in the set of all artifacts. Artifact
resolution should take place separately for project and plugin artifacts anyway
because the list of remote repositories may differ for each. Each subproject may
also have different remote repositories, so we should not do all artifact
resolution in one project scope.

The set of resolved artifacts is now a TreeSet instead of a HashSet because the
latter tends to assume that object equality is object identity, i.e. that two
objects are only equal if they are the same object (in memory) which is not what
we want.