Enter only required information into manifest
The manifest need only record the download url, SHA-1 checksum, and destination
path (in the local repository). All other fields are removed and NixArtifact is
renamed to ManifestEntry.

A warning is produced for all artifacts that cannot resolved to a remote
repository (for download) or to a workspace repository (in which case, no
manifest entry is emitted).